basicInstagram/pyproject.toml
[project]
name = "basicinstagram"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.128.0",
    "fastapi-users[sqlalchemy]>=15.0.3",
    "imagekitio>=5.1.1",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.40.0",
]
